+++ src/core/dataclass.cpp
@@ -860,15 +860,7 @@ void DataClass::SwitchConentrations()
     }
 }
 
-qreal DataClass::InitialGuestConcentration(int i) const
-{
-    return d->m_independent_model->data(!HostAssignment(), i) * d->m_scaling[!HostAssignment()];
-}
 
-qreal DataClass::InitialHostConcentration(int i) const
-{
-    return d->m_independent_model->data(HostAssignment(), i) * d->m_scaling[HostAssignment()];
-}
 
 const QJsonObject DataClass::ExportData() const
 {